protect user 1 during suppression of blocked user

+++ scripts/sanitize.sql
@@ -20,7 +20,7 @@ DELETE FROM variable WHERE name = 'cron_semaphore';
 -- Get rid of unpublished/blocked nodes, users, comments and related data in other tables.
 DELETE FROM node WHERE status <> 1;
 DELETE FROM comments WHERE status <> 0;
-DELETE FROM users WHERE status <> 1 AND uid <> 0;
+DELETE FROM users WHERE status <> 1 AND (uid <> 0 AND uid <> 1);
 DELETE node FROM node LEFT JOIN users ON node.uid = users.uid WHERE users.uid IS NULL;
 DELETE node_revisions FROM node_revisions LEFT JOIN node ON node.nid = node_revisions.nid WHERE node.nid IS NULL;
 DELETE comments FROM comments LEFT JOIN node ON node.nid = comments.nid WHERE node.nid IS NULL;
